Add user to log controller

+++ controllers/logs.go
@@ -6,6 +6,7 @@ import (
 	"strconv"
 
 	"github.com/antonve/logger-api/models"
+	jwt "github.com/dgrijalva/jwt-go"
 
 	"github.com/labstack/echo"
 )
@@ -39,7 +40,8 @@ func APILogsPost(context echo.Context) error {
 // APILogsGetAll gets all logs
 func APILogsGetAll(context echo.Context) error {
 	logCollection := models.LogCollection{Logs: make([]models.Log, 0)}
-	err := logCollection.GetAll()
+	user := context.Get("user").(*jwt.Token).Claims.(*models.JwtClaims).User
+	err := logCollection.GetAllFromUser(user.ID)
 
 	if err != nil {
 		return Return500(context, err)
@@ -63,7 +65,12 @@ func APILogsGetByID(context echo.Context) error {
 	}
 
 	if log == nil {
-		return Return404(context, fmt.Errorf("No Log found with id %v", id))
+		return Return404(context, fmt.Errorf("no Log found with id %v", id))
+	}
+
+	user := context.Get("user").(*jwt.Token).Claims.(*models.JwtClaims).User
+	if log.IsOwner(user.ID) {
+		return Return403(context, fmt.Errorf("log entry doesn't belong to user"))
 	}
 
 	return context.JSON(http.StatusOK, log)
@@ -84,10 +91,19 @@ func APILogsUpdate(context echo.Context) error {
 	if err != nil {
 		return Return500(context, err)
 	}
-	log.ID = id
 
 	// Update
 	logCollection := models.LogCollection{}
+	currentLog, err := logCollection.Get(id)
+	if err != nil {
+		return Return500(context, err)
+	}
+
+	user := context.Get("user").(*jwt.Token).Claims.(*models.JwtClaims).User
+	if !currentLog.IsOwner(user.ID) {
+		return Return403(context, fmt.Errorf("log entry doesn't belong to user"))
+	}
+
 	err = logCollection.Update(log)
 	if err != nil {
 		return Return500(context, err)
@@ -109,6 +125,16 @@ func APILogsDelete(context echo.Context) error {
 	log.ID = id
 
 	logCollection := models.LogCollection{}
+	currentLog, err := logCollection.Get(id)
+	if err != nil {
+		return Return500(context, err)
+	}
+
+	user := context.Get("user").(*jwt.Token).Claims.(*models.JwtClaims).User
+	if !currentLog.IsOwner(user.ID) {
+		return Return403(context, fmt.Errorf("log entry doesn't belong to user"))
+	}
+
 	err = logCollection.Delete(log)
 	if err != nil {
 		return Return500(context, err)
